Abstract custom icon pack logic from icon pack class
